Move 'overload' types inline.
